requirements.txt
ExifRead~=3.0.0
beautifulsoup4~=4.11.2
soupsieve~=2.3.2.post1
opencv-python~=4.7.0.68
numpy~=1.24.1
defusedxml~=0.7.1
cffi~=1.15.1
Pillow~=9.4.0
ipython~=8.9.0
jupyter~=1.0.0
pip~=22.3.1
attrs~=22.2.0
wheel~=0.38.4
tornado~=6.2
Jinja2~=3.1.2
pyzmq~=25.0.0
pexpect~=4.8.0
comm~=0.1.2
traitlets~=5.9.0
fqdn~=1.5.1
parso~=0.8.3
jedi~=0.18.2
pytz~=2022.7.1
QtPy~=2.3.0
packaging~=23.0
requests~=2.28.2
ipywidgets~=8.0.4
notebook~=6.5.2
PyYAML~=6.0
anyio~=3.6.2
sniffio~=1.3.0
arrow~=1.2.3
python-dateutil~=2.8.2
setuptools~=65.5.1
scipy~=1.10.0
matplotlib~=3.6.3
six~=1.16.0
webencodings~=0.5.1
bleach~=6.0.0
tinycss2~=1.2.1
MarkupSafe~=2.1.2
joblib~=1.2.0
psutil~=5.9.4
pandas~=1.5.3
certifi~=2022.12.7
debugpy~=1.6.6
decorator~=5.1.1
Pygments~=2.14.0
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.6
mistune~=2.0.4
ptyprocess~=0.7.0
scikit-learn~=1.2.1
threadpoolctl~=3.1.0
nbformat~=5.7.3
nbclient~=0.7.2
nbconvert~=7.2.9
fastjsonschema~=2.16.2
jsonschema~=4.17.3
Send2Trash~=1.8.0
terminado~=0.17.1
nbclassic~=0.5.1
idna~=3.4
urllib3~=1.26.14
asttokens~=2.2.1
contourpy~=1.0.7
executing~=1.2.0
fonttools~=4.38.0
ipykernel~=6.21.1
pandocfilters~=1.5.0
pycparser~=2.21
pyparsing~=3.0.9
qtconsole~=5.4.0
websocket-client~=1.5.0
pyrsistent~=0.19.3
webcolors~=1.12
jsonpointer~=2.3
isoduration~=20.11.0
cycler~=0.11.0
kiwisolver~=1.4.4
zipp~=3.12.0
platformdirs~=2.6.2
tqdm~=4.64.1